feat: use isReflectNil instead of isEmptyValue to gate Transformers
Mergo handles merges with zero-value fields as follows:

	1. dst: true       src: false      ->     dst: true
	2. dst: nil        src: false      ->     dst: false
	3. dst: nil	   src: true	   ->	  dst: true
	4. dst: false	   src: true 	   ->	  dst: true

In some cases (https://github.com/imdario/mergo/issues/166) it may be desirable
to change this default behaviour using a custom Transformer, for example
to not overwrite 'dst' when it's already set to 'false' (or '0' etc).
Unfortunately, in such cases (when dst holds a zero-value) no Transformers can
be applied due to the way these are gated (https://github.com/imdario/mergo/issues/166).

Here we are changing how transformers are gated by using the isReflectNil function
insted of isEmptyValue.

Note that:
- this change doesn't affect the default behaviour of mergo, as it only affects
  how transformers are gated
- since custom transformers need to match data types in order to execute,
  this shouldn't affect any existing transformers either